+++ lib/src/game.dart
@@ -4,19 +4,43 @@ library;
 import 'package:scopa_lib/scopa_lib.dart';
 import 'package:scopa_lib/tabletop_lib.dart';
 
-/// Driving class for the Scopa game
+const Map<int, int> _primeValues = {
+  7: 21,
+  6: 18,
+  1: 16,
+  5: 15,
+  4: 14,
+  3: 13,
+  2: 12,
+  8: 10,
+  9: 10,
+  10: 10
+};
+
+/// Driving class for the Scopa game.
 class Game {
   final manager = HandManager(ScopaDeck.instance);
   final List<Team> _teams;
+  final _teamScores = <Team, int>{};
   late final ScopaTable table;
 
+  /// The list of [Team]s in the game.
   List<Team> get teams => List.unmodifiable(_teams);
+
+  /// The scores of each team [Team].
+  Map<Team, int> get teamScores => Map.unmodifiable(_teamScores);
+
+  /// A map of [Player]s to which [Team] they're in.
+  Map<Player, Team> get playerTeams => Map.fromEntries(_teams
+      .expand((team) => team.players.map((player) => MapEntry(player, team))));
 
   Game(this._teams) {
     if (_teams.isEmpty) {
       table = ScopaTable(0, manager);
       return;
     }
+
+    _teamScores.addEntries(_teams.map((team) => MapEntry(team, 0)));
 
     final numPlayers = _teams.fold(
         0, (previousValue, element) => previousValue += element.players.length);
@@ -42,5 +66,147 @@ class Game {
     round.dealPlayers();
     round.dealRound();
     return round;
+  }
+
+  Iterable<Team>? scoreRound(ScopaRound round) {
+    // Add points for player scopas
+    for (final team in _teams) {
+      final oldScore = _teamScores[team]!;
+      final numScopas = team.players.fold(
+          0, (previousValue, player) => round.scopas[player]! + previousValue);
+      _teamScores[team] = oldScore + numScopas;
+    }
+
+    // Add a point to the player with the most fishes
+    var isTied = false;
+    final mostFish = round.captureHands.entries
+        .fold<MapEntry<Player, Hand>?>(null, (previousValue, element) {
+      if (previousValue == null) return element;
+
+      final currentIsGreater =
+          element.value.cards.length > previousValue.value.cards.length;
+      final currentIsEqual =
+          element.value.cards.length == previousValue.value.cards.length;
+
+      if (currentIsEqual) {
+        isTied = true;
+      }
+
+      if (currentIsGreater) {
+        isTied = false;
+      }
+
+      return currentIsGreater ? element : previousValue;
+    })!.key;
+
+    if (isTied == false) {
+      final score = _teamScores[playerTeams[mostFish]]!;
+      _teamScores[playerTeams[mostFish]!] = score + 1;
+    }
+
+    // Add a point for capturing the most coppes
+    final coppeHands = <Player, int>{};
+    for (final player in playerTeams.keys) {
+      final fishedCards = round.captureHands[player]!.cards;
+      if (fishedCards.any((card) => card.suite == 'Coppe')) {
+        final numCoppes = fishedCards.fold(
+            0,
+            (previousValue, element) =>
+                element.suite == 'Coppe' ? previousValue + 1 : previousValue);
+
+        coppeHands[player] = numCoppes;
+      }
+    }
+
+    isTied = false;
+    final mostCoppes = coppeHands.entries.fold(0, (previousValue, element) {
+      final currentIsGreater = element.value > previousValue;
+      final currentIsEqual = element.value == previousValue;
+
+      if (currentIsEqual) {
+        isTied = true;
+      }
+
+      if (currentIsGreater) {
+        isTied = false;
+      }
+
+      return currentIsGreater ? element.value : previousValue;
+    });
+
+    if (isTied == false && coppeHands.isNotEmpty) {
+      final player = coppeHands.entries
+          .firstWhere((element) => element.value == mostCoppes)
+          .key;
+
+      final oldScore = _teamScores[playerTeams[player]]!;
+      _teamScores[playerTeams[player]!] = oldScore + 1;
+    }
+
+    // Add point for fishing the 7 coppe
+    containsCoppe7(element) => element.value.cards.contains(Card('Coppe', 7));
+
+    if (round.captureHands.entries.any((element) => containsCoppe7(element))) {
+      final player = round.captureHands.entries
+          .singleWhere((element) => containsCoppe7(element))
+          .key;
+
+      final oldScore = _teamScores[playerTeams[player]]!;
+      _teamScores[playerTeams[player]!] = oldScore + 1;
+    }
+
+    // Add point for fishing the highest prime
+    final primeScores = <Player, int>{};
+    for (final player in playerTeams.keys) {
+      primeScores[player] = manager.deck.suites.fold(
+          0,
+          (playerPrimeScore, suite) => round.captureHands[player]!
+              .getSuiteCards(suite)
+              .fold(
+                  0,
+                  (suiteHighestPrimeValue, card) =>
+                      _primeValues[card.value]! > suiteHighestPrimeValue
+                          ? _primeValues[card.value]!
+                          : suiteHighestPrimeValue));
+    }
+
+    final highestPrimeScorePlayer = primeScores.entries.fold<Player?>(null,
+        (previousValue, playerPrimeScore) {
+      if (previousValue == null) return playerPrimeScore.key;
+
+      if (playerPrimeScore.value == primeScores[previousValue]) {
+        isTied = true;
+        return previousValue;
+      }
+
+      if (playerPrimeScore.value > primeScores[previousValue]!) {
+        isTied = false;
+        return playerPrimeScore.key;
+      }
+
+      return previousValue;
+    });
+
+    if (highestPrimeScorePlayer != null && isTied == false) {
+      final team = playerTeams[highestPrimeScorePlayer]!;
+      final oldScore = teamScores[team]!;
+      _teamScores[team] = oldScore + 1;
+    }
+
+    // Return any winning teams
+    final winningTeams =
+        teamScores.entries.where((element) => element.value >= 11);
+
+    if (winningTeams.isNotEmpty) {
+      final highestScore = winningTeams.fold(
+          0,
+          (previousValue, element) =>
+              element.value > previousValue ? element.value : previousValue);
+      return winningTeams
+          .where((element) => element.value == highestScore)
+          .map((e) => e.key);
+    }
+
+    return null;
   }
 }