+++ test/scopa_round_test.dart
@@ -118,11 +118,35 @@ void main() {
         round.resetPool();
         round.dealPlayers();
 
-        for (var i = 0; i > 3; i++) {
+        for (var i = 0; i < 3; i++) {
           round.play(playerHand.cards[playerHand.cards.length - 1]);
         }
 
         expect(playerHand.cards.length, equals(3));
+      });
+
+      test('does not redeal players if there are not enough pool cards', () {
+        final roundStuff = getTestRound(2);
+        final round = roundStuff.$1;
+        final manager = roundStuff.$2;
+        final table = roundStuff.$3;
+
+        for (var i = 0; i < 8; i++) {
+          manager.deal(manager.deck.cards.elementAt(i), table.pool);
+        }
+
+        round.dealPlayers();
+        assert(table.pool.cards.length == 2);
+
+        bool? roundState;
+        for (var i = 0; i < 6; i++) {
+          final playerHand = round.playerHands[round.currentPlayer]!;
+          roundState =
+              round.play(playerHand.cards[playerHand.cards.length - 1]);
+        }
+
+        assert(roundState == false);
+        expect(table.pool.cards.length, equals(2));
       });
 
       test('sets the next player if not ending', () {
@@ -196,22 +220,35 @@ void main() {
         expect(round.captureHands.values.first.cards, isEmpty);
       });
 
+      test('adds a scopa when a scopa is scored', () {
+        final round = getTestRound();
+        round.$1.resetPool();
+        final playCard = Card('Bastoni', 6);
+        final matchCards = [Card('Coppe', 2), Card('Denari', 4)];
+        round.$2.dealAll(matchCards, round.$3.round);
+        round.$2.deal(playCard, round.$1.playerHands.values.first);
+
+        round.$1.play(playCard, matchCards);
+
+        expect(round.$1.scopas[round.$1.currentPlayer], equals(1));
+      });
+
+      test('redeals the round hand when empty', () {
+        final round = getTestRound();
+        round.$1.resetPool();
+        final playCard = Card('Bastoni', 6);
+        final matchCards = [Card('Coppe', 2), Card('Denari', 4)];
+        round.$2.dealAll(matchCards, round.$3.round);
+        round.$2.deal(playCard, round.$1.playerHands.values.first);
+
+        round.$1.play(playCard, matchCards);
+
+        expect(round.$3.round.cards, hasLength(4));
+      });
+
       group('returns', () {
-        test('a scopa result if a scopa was scored', () {
-          final round = getTestRound();
-          round.$1.resetPool();
-          final playCard = Card('Bastoni', 6);
-          final matchCards = [Card('Coppe', 2), Card('Denari', 4)];
-          round.$2.dealAll(matchCards, round.$3.round);
-          round.$2.deal(playCard, round.$1.playerHands.values.first);
-
-          final roundState = round.$1.play(playCard, matchCards);
-
-          expect(roundState, equals(RoundState.scopa));
-        });
-
         // TODO: Update test to end when all player hands are empty too
-        test('an ending result when the pool hand is empty', () {
+        test('false when the pool hand is empty', () {
           final roundInfo = getTestRound();
           final round = roundInfo.$1;
           final manager = roundInfo.$2;
@@ -222,13 +259,13 @@ void main() {
 
           final roundState = round.play(Card('Denari', 2), [Card('Coppe', 2)]);
 
-          expect(roundState, equals(RoundState.ending));
+          expect(roundState, equals(false));
         });
 
-        test('an ending result when there are no players', () {
+        test('false when there are no players', () {
           final round = getTestRound(0).$1;
 
-          expect(round.play(Card('Denari', 2)), equals(RoundState.ending));
+          expect(round.play(Card('Denari', 2)), equals(false));
         });
       });
     });